examples/pyproject.toml
[project]
name = "microtorch-examples"
version = "0.1.0"
description = "Example neural networks written with MicroTorch"
readme = "README.md"
authors = [
    { name = "Rafid", email = "rafidka@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "microtorch",
    "pillow>=11.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
microtorch = { path = "../", editable = true }

[project.scripts]
mnist = "microtorch_examples.mnist:main"
